Add Container around `main` & `aside` content
Add a container element to wrap around the `main` and `aside` into
applied styles that related to both of them and neither to the page
header or footer. This allows `flex` and `width` to be applied to `main`
and `aside` to bring it closer to the appearance of the mockup example.